* install menu item in /usr/local/etc/ext/$area * add do you want to install this package and do you want to delete ... * add needed text to system to run the package.  ex: add text to /etc/crontab

+++ usr/local/www/pkg_mgr_installed.php
@@ -85,8 +85,7 @@ require("xmlparse_pkg.inc");
                                     <?= $pkg['descr'] ?>
                                 </td>
                                 <td valign="middle" class="list" nowrap>
-                                    <a href="pkg_mgr_delete.php?id=<?=$i;?>"><img src="x.gif" width="17" height="17" border="0"></a>
-
+                                    <a onclick="return confirm('Do you really want to remove this package?')" href="pkg_mgr_delete.php?id=<?=$i;?>"><img src="x.gif" width="17" height="17" border="0"></a>
                                 </td>
                             </tr>
